factor out some update loop logic into separate package
