+++ wfupdater/sfn_events.go
@@ -1,11 +1,12 @@
-package executor
+package wfupdater
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/sfn"
 )
 
-func reverseHistory(events []*sfn.HistoryEvent) []*sfn.HistoryEvent {
+// ReverseHistory reverses the order of the events in the history.
+func ReverseHistory(events []*sfn.HistoryEvent) []*sfn.HistoryEvent {
 	numEvents := len(events)
 	reversedEvents := make([]*sfn.HistoryEvent, numEvents)
 	for i := 0; i < numEvents; i++ {
